requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
pandas==2.2.1
httpx==0.28.1